# use __ARM_NEON along with __ARM_NEON__ to detect NEON availability