@mittyorz 2021-09-05
### 600MB(5分間)のTSを用いたベンチマーク

復号化ファイルはバイナリ比較し違いがないことを確認ずみです

#### without NEON

```
 Performance counter stats for 'b25':

         10,579.27 msec task-clock                #    1.000 CPUs utilized
                84      context-switches          #    7.940 /sec
                 0      cpu-migrations            #    0.000 /sec
               457      page-faults               #   43.198 /sec
    15,867,812,930      cycles                    #    1.500 GHz
    16,824,729,091      instructions              #    1.06  insn per cycle
   <not supported>      branches
         4,816,977      branch-misses

      10.584478676 seconds time elapsed

       9.844824000 seconds user
       0.708124000 seconds sys
```

#### with NEON

```
 Performance counter stats for 'b25':

          6,127.06 msec task-clock                #    1.000 CPUs utilized
                16      context-switches          #    2.611 /sec
                 0      cpu-migrations            #    0.000 /sec
               459      page-faults               #   74.914 /sec
     9,190,401,336      cycles                    #    1.500 GHz
     6,756,342,641      instructions              #    0.74  insn per cycle
   <not supported>      branches
         1,626,470      branch-misses

       6.128483999 seconds time elapsed

       5.522691000 seconds user
       0.589216000 seconds sys
```

---

ファイルの復号とは別に、PX-Q3PE4/mirakurun/chinachuを用いて4ch同時復号化でRasPiのCPU負荷・消費電力を比較してみましたが、誤差と思われるレベルの違いしか発生しませんでした。